Update Get SellerOrder By Id response

+++ WebApi/Features/Users/DeactivateUserByUserId.cs
@@ -145,7 +145,7 @@ public class DeactivateUserByUserId : ControllerBase
 
                 var customerWallet = await context.Wallets.FirstOrDefaultAsync(w => w.Id == sellerOrderDetail.Order.WalletTracking.WalletId);
 
-                int totalAmount = 0;
+                long totalAmount = 0;
 
                 WalletTracking walletTracking = new WalletTracking()
                 {
@@ -163,7 +163,7 @@ public class DeactivateUserByUserId : ControllerBase
                 foreach (var soi in selelrOrderItems)
                 {
                     int discountPercentage = soi.GadgetDiscount != null ? soi.GadgetDiscount.DiscountPercentage : 0;
-                    totalAmount += soi.GadgetQuantity * (int)Math.Ceiling(soi.GadgetPrice * (1 - discountPercentage / 100.0));
+                    totalAmount += soi.GadgetQuantity * (long)Math.Ceiling(soi.GadgetPrice * (1 - discountPercentage / 100.0));
                     soi.Gadget.Quantity += soi.GadgetQuantity;
                 }
